Move logger creation to separate method
-Add conditional Seq logging
